requirements.txt
fastapi==0.112.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
python-dotenv==1.0.1
pydantic==2.6.3
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0